+++ scenedetect/common.py
@@ -66,6 +66,9 @@ import typing as ty
 from dataclasses import dataclass
 from fractions import Fraction
 
+
+_USE_PTS_IN_DEVELOPMENT = False
+
 ##
 ## Type Aliases
 ##