+++ scenedetect/detectors/content_detector.py
@@ -22,6 +22,7 @@ from dataclasses import dataclass
 import cv2
 import numpy
 
+from scenedetect.common import FrameTimecode
 from scenedetect.detector import FlashFilter, SceneDetector
 
 
@@ -125,7 +126,6 @@ class ContentDetector(SceneDetector):
         """
         super().__init__()
         self._threshold: float = threshold
-        self._min_scene_len: int = min_scene_len
         self._last_above_threshold: ty.Optional[int] = None
         self._last_frame: ty.Optional[ContentDetector._FrameData] = None
         self._weights: ContentDetector.Components = weights
@@ -137,12 +137,13 @@ class ContentDetector(SceneDetector):
                 raise ValueError("kernel_size must be odd integer >= 3")
             self._kernel = numpy.ones((kernel_size, kernel_size), numpy.uint8)
         self._frame_score: ty.Optional[float] = None
+        # TODO(v0.7): Handle timecodes in filter.
         self._flash_filter = FlashFilter(mode=filter_mode, length=min_scene_len)
 
     def get_metrics(self):
         return ContentDetector.METRIC_KEYS
 
-    def _calculate_frame_score(self, frame_num: int, frame_img: numpy.ndarray) -> float:
+    def _calculate_frame_score(self, timecode: FrameTimecode, frame_img: numpy.ndarray) -> float:
         """Calculate score representing relative amount of motion in `frame_img` compared to
         the last time the function was called (returns 0.0 on the first call)."""
         # TODO: Add option to enable motion estimation before calculating score components.
@@ -178,13 +179,15 @@ class ContentDetector(SceneDetector):
         if self.stats_manager is not None:
             metrics = {self.FRAME_SCORE_KEY: frame_score}
             metrics.update(score_components._asdict())
-            self.stats_manager.set_metrics(frame_num, metrics)
+            self.stats_manager.set_metrics(timecode.frame_num, metrics)
 
         # Store all data required to calculate the next frame's score.
         self._last_frame = ContentDetector._FrameData(hue, sat, lum, edges)
         return frame_score
 
-    def process_frame(self, frame_num: int, frame_img: numpy.ndarray) -> ty.List[int]:
+    def process_frame(
+        self, timecode: FrameTimecode, frame_img: numpy.ndarray
+    ) -> ty.List[FrameTimecode]:
         """Process the next frame. `frame_num` is assumed to be sequential.
 
         Args:
@@ -196,12 +199,12 @@ class ContentDetector(SceneDetector):
            ty.List[int]: List of frames where scene cuts have been detected. There may be 0
             or more frames in the list, and not necessarily the same as frame_num.
         """
-        self._frame_score = self._calculate_frame_score(frame_num, frame_img)
+        self._frame_score = self._calculate_frame_score(timecode, frame_img)
         if self._frame_score is None:
             return []
 
         above_threshold: bool = self._frame_score >= self._threshold
-        return self._flash_filter.filter(frame_num=frame_num, above_threshold=above_threshold)
+        return self._flash_filter.filter(timecode=timecode, above_threshold=above_threshold)
 
     def _detect_edges(self, lum: numpy.ndarray) -> numpy.ndarray:
         """Detect edges using the luma channel of a frame.